requirements.txt
annotated-types==0.6.0
arabic-reshaper==3.0.0
asgiref==3.8.1
asn1crypto==1.5.1
certifi==2024.2.2
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
crispy-bootstrap4==2024.1
cryptography==42.0.7
cssselect2==0.7.0
defusedxml==0.7.1
diff-match-patch==20230430
Django==5.0.4
django-annoying==0.10.6
django-crispy-forms==2.1
django-extensions==3.2.3
django-import-export==4.0.1
django-qr-code==4.0.1
et-xmlfile==1.1.0
html5lib==1.1
idna==3.7
lxml==5.2.1
MarkupPy==1.14
mysqlclient==2.2.4
numpy==1.26.4
odfpy==1.4.1
openpyxl==3.1.2
oscrypto==1.3.0
pandas==2.2.2
pillow==10.3.0
pycparser==2.22
pydantic==2.7.1
pydantic_core==2.18.2
pyHanko==0.25.0
pyhanko-certvalidator==0.26.3
pypdf==4.2.0
pypng==0.20220715.0
python-bidi==0.4.2
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
qrcode==7.4.2
reportlab==4.0.9
requests==2.31.0
segno==1.6.1
six==1.16.0
sqlparse==0.5.0
svglib==1.5.1
tablib==3.5.0
tabulate==0.9.0
tinycss2==1.3.0
typing_extensions==4.11.0
tzdata==2024.1
tzlocal==5.2
uritools==4.0.2
urllib3==2.2.1
webencodings==0.5.1
xhtml2pdf==0.2.15
xlrd==2.0.1
xlwt==1.3.0
